requirements.txt
django
django-bootstrap3
django-registration
django-phonenumber-field

# for fixtures
pyyaml
